Move Client functions that depend on an Ecto repo to a separate module
